osd: move Session class out of OSD
